LPS-22792 Move logic of a category being required to the model
